버킷상세 > 달성횟수 변경 api 연동

+++ app/src/main/java/com/zinc/berrybucket/ui/presentation/detail/component/GoalCountUpdateDialog.kt
@@ -39,7 +39,7 @@ fun GoalCountUpdateDialog(
 
                 is SingleTextFieldDialogEvent.Update -> event.invoke(
                     GoalCountUpdateEvent.CountUpdate(
-                        updateGoalCount
+                        updateGoalCount.toInt()
                     )
                 )
             }
